refactor: move generate_paradigm to paradigm.generation
