+++ CreeDictionary/CreeDictionary/paradigm/generation.py
